+++ external/vulkancts/modules/vulkan/dynamic_state/vktDynamicStateBaseClass.cpp
@@ -142,7 +142,7 @@ void DynamicStateBaseClass::initialize (void)
 											m_context.getDefaultAllocator(), vk::MemoryRequirement::HostVisible);
 
 	deUint8* ptr = reinterpret_cast<unsigned char *>(m_vertexBuffer->getBoundMemory().getHostPtr());
-	deMemcpy(ptr, &m_data[0], dataSize);
+	deMemcpy(ptr, &m_data[0], (size_t)dataSize);
 
 	vk::flushMappedMemoryRange(m_vk, device,
 		m_vertexBuffer->getBoundMemory().getMemory(),
